checkLoaded should use getAttribute
When a script is asked for it's src property it returns the full path of the loaded script. 
If some script is aliased using a relative path, and also preloaded in the page (for example jQuery), it will be reloaded whenever the alias gets its first hit, since the checkLoaded has the full path version as loaded, and not the relative version.

This fixes that issue.